+++ src/main/scala/cognite/spark/v1/AssetsRelation.scala
@@ -6,7 +6,7 @@ import cats.effect.IO
 import cats.implicits._
 import cognite.spark.v1.PushdownUtilities._
 import cognite.spark.v1.SparkSchemaHelper._
-import com.cognite.sdk.scala.common.{WithExternalId, WithId}
+import com.cognite.sdk.scala.common._
 import com.cognite.sdk.scala.v1.resources.Assets
 import com.cognite.sdk.scala.v1._
 import fs2.Stream
@@ -16,7 +16,7 @@ import org.apache.spark.sql.types._
 import org.apache.spark.sql.{Row, SQLContext}
 
 class AssetsRelation(config: RelationConfig)(val sqlContext: SQLContext)
-    extends SdkV1Relation[Asset, Long](config, "assets")
+    extends SdkV1Relation[AssetsReadSchema, Long](config, "assets")
     with InsertableRelation
     with WritableRelation {
   import CdpConnector._
@@ -24,8 +24,8 @@ class AssetsRelation(config: RelationConfig)(val sqlContext: SQLContext)
   override def getStreams(filters: Array[Filter])(
       client: GenericClient[IO],
       limit: Option[Int],
-      numPartitions: Int): Seq[Stream[IO, Asset]] = {
-    val fieldNames = Array("name", "source", "dataSetId")
+      numPartitions: Int): Seq[Stream[IO, AssetsReadSchema]] = {
+    val fieldNames = Array("name", "source", "dataSetId", "labels")
     val pushdownFilterExpression = toPushdownFilterExpression(filters)
     val getAll = shouldGetAll(pushdownFilterExpression, fieldNames)
     val params = pushdownToParameters(pushdownFilterExpression)
@@ -44,18 +44,29 @@ class AssetsRelation(config: RelationConfig)(val sqlContext: SQLContext)
     // Merge streams related to each partition to make sure duplicate values are read into
     // the same RDD partition
     streamsPerFilter.transpose
-      .map(s => s.reduce(_.merge(_)))
+      .map(
+        s =>
+          s.reduce(_.merge(_))
+            .map(
+              _.into[AssetsReadSchema]
+                .withFieldComputed(_.labels, u => cogniteExternalIdSeqToStringSeq(u.labels))
+                .transform))
   }
 
   private def assetsFilterFromMap(m: Map[String, String]): AssetsFilter =
     AssetsFilter(
       name = m.get("name"),
       source = m.get("source"),
-      dataSetIds = m.get("dataSetId").map(assetIdsFromWrappedArray(_).map(CogniteInternalId))
+      dataSetIds = m.get("dataSetId").map(idsFromWrappedArray(_).map(CogniteInternalId)),
+      labels = m.get("labels").map(externalIdsToContainsAny).getOrElse(None)
     )
 
   override def insert(rows: Seq[Row]): IO[Unit] = {
-    val assets = rows.map(fromRow[AssetCreate](_))
+    val assetsInsertions = rows.map(fromRow[AssetsInsertSchema](_))
+    val assets = assetsInsertions.map(
+      _.into[AssetCreate]
+        .withFieldComputed(_.labels, u => stringSeqToCogniteExternalIdSeq(u.labels))
+        .transform)
     client.assets
       .create(assets)
       .flatTap(_ => incMetrics(itemsCreated, assets.size)) *> IO.unit
@@ -78,35 +89,49 @@ class AssetsRelation(config: RelationConfig)(val sqlContext: SQLContext)
   }
 
   override def upsert(rows: Seq[Row]): IO[Unit] = {
+    logger.info("We're upserting")
     val assets = rows.map(fromRow[AssetsUpsertSchema](_))
     val (itemsToUpdate, itemsToUpdateOrCreate) =
       assets.partition(r => r.id.exists(_ > 0) || (r.name.isEmpty && r.externalId.nonEmpty))
 
+    logger.info(
+      s"itemsToUpdate: ${itemsToUpdate.length}, itemsToUpdateOrCreate: ${itemsToUpdateOrCreate.length}")
     if (itemsToUpdateOrCreate.exists(_.name.isEmpty)) {
       throw new CdfSparkIllegalArgumentException("The name field must be set when creating assets.")
     }
 
     genericUpsert[Asset, AssetsUpsertSchema, AssetCreate, AssetUpdate, Assets[IO]](
       itemsToUpdate,
-      itemsToUpdateOrCreate.map(_.into[AssetCreate].withFieldComputed(_.name, _.name.get).transform),
+      itemsToUpdateOrCreate.map(
+        _.into[AssetCreate]
+          .withFieldComputed(_.name, _.name.get)
+          .withFieldComputed(_.labels, u => stringSeqToCogniteExternalIdSeq(u.labels))
+          .transform),
       isUpdateEmpty,
-      client.assets)
+      client.assets
+    )
   }
 
   override def getFromRowsAndCreate(rows: Seq[Row], doUpsert: Boolean = true): IO[Unit] = {
-    val assets = rows.map(fromRow[AssetCreate](_))
+    val assetsUpserts = rows.map(fromRow[AssetsUpsertSchema](_))
+    val assets = assetsUpserts.map(
+      _.into[AssetCreate]
+        .withFieldComputed(_.name, u => u.name.get)
+        .withFieldComputed(_.labels, u => stringSeqToCogniteExternalIdSeq(u.labels))
+        .transform)
     createOrUpdateByExternalId[Asset, AssetUpdate, AssetCreate, Assets[IO]](
       Set.empty,
       assets,
       client.assets,
       doUpsert = true)
+
   }
 
-  override def schema: StructType = structType[Asset]
+  override def schema: StructType = structType[AssetsReadSchema]
 
-  override def toRow(a: Asset): Row = asRow(a)
+  override def toRow(a: AssetsReadSchema): Row = asRow(a)
 
-  override def uniqueId(a: Asset): Long = a.id
+  override def uniqueId(a: AssetsReadSchema): Long = a.id
 }
 
 object AssetsRelation extends UpsertSchema {
@@ -124,7 +149,8 @@ final case class AssetsUpsertSchema(
     metadata: Option[Map[String, String]] = None,
     parentId: Option[Long] = None,
     parentExternalId: Option[String] = None,
-    dataSetId: Option[Long] = None
+    dataSetId: Option[Long] = None,
+    labels: Option[Seq[String]] = None
 ) extends WithExternalId
     with WithId[Option[Long]]
 
@@ -136,20 +162,23 @@ final case class AssetsInsertSchema(
     externalId: Option[String] = None,
     metadata: Option[Map[String, String]] = None,
     parentExternalId: Option[String] = None,
-    dataSetId: Option[Long] = None
+    dataSetId: Option[Long] = None,
+    labels: Option[Seq[String]] = None
 )
 
 final case class AssetsReadSchema(
     externalId: Option[String] = None,
     name: String,
     parentId: Option[Long] = None,
+    parentExternalId: Option[String] = None,
     description: Option[String] = None,
     metadata: Option[Map[String, String]] = None,
     source: Option[String] = None,
     id: Long = 0,
     createdTime: Instant = Instant.ofEpochMilli(0),
     lastUpdatedTime: Instant = Instant.ofEpochMilli(0),
-    rootId: Long = 0,
+    rootId: Option[Long] = Some(0),
     aggregates: Option[Map[String, Long]] = None,
-    dataSetId: Option[Long] = None
+    dataSetId: Option[Long] = None,
+    labels: Option[Seq[String]] = None
 )